+++ src/components/Select/types.ts
@@ -61,12 +61,15 @@ export type SelectProps<T = any> = QAProps &
             | React.ReactElement<SelectOptionGroup<T>, typeof OptionGroup>[];
     };
 
-export type SelectOption<T = any> = ControlGroupOption & {
-    text?: string;
-    data?: T;
-};
+export type SelectOption<T = any> = QAProps &
+    ControlGroupOption & {
+        text?: string;
+        data?: T;
+    };
 
 export type SelectOptionGroup<T = any> = {
+    /** Label is a string which displayed above the options group.
+     * If label is empty string, group item height will be 0 and only border will be displayed */
     label: string;
     options?: SelectOption<T>[];
     children?: